Deprecate dtype

+++ src/metrics/mse/script.py
@@ -18,8 +18,8 @@ print("Load data", flush=True)
 input_denoised = ad.read_h5ad(par['input_prediction'])
 input_test = ad.read_h5ad(par['input_test'])
 
-test_data = ad.AnnData(X=input_test.layers["counts"], dtype="float")
-denoised_data = ad.AnnData(X=input_denoised.layers["denoised"], dtype="float")
+test_data = ad.AnnData(X=input_test.layers["counts"])
+denoised_data = ad.AnnData(X=input_denoised.layers["denoised"])
 
 print("Normalize data", flush=True)
